% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fold.R
\name{foldCode}
\alias{foldCode}
\title{Fold a file}
\usage{
foldCode(contents = NA)
}
\arguments{
\item{contents}{the code to be folded; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}
}
\value{
The folded code in a character string.
}
\description{
Fold any file (wrap lines to 80 characters). The result is
  not indented but the advantage is that this works for big files.
}
\note{
This function requires the command line utility \code{fold}. You can
  find it in Rtools.
}
