% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getPrettifiableLanguages}
\alias{getPrettifiableLanguages}
\title{Prettifiable languages}
\usage{
getPrettifiableLanguages()
}
\description{
Returns the list of languages that are supported by this
  package.
}
