% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyAddins-imports.R
\docType{import}
\name{prettifyAddins-imports}
\alias{prettifyAddins-imports}
\alias{install_phantomjs}
\title{Install PhantomJS}
\description{
This function is imported from the 'webdriver' package.
  Follow the link to its documentation:
  \code{\link[webdriver:install_phantomjs]{install_phantomjs}}
}
