% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyCLANG.R
\name{prettifyCLANG}
\alias{prettifyCLANG}
\title{Prettify C, C++, Java}
\usage{
prettifyCLANG(contents = NA, language = NA, tabSize = NULL)
}
\arguments{
\item{contents}{the code to be prettified; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}

\item{language}{the language of the code; when the contents is read from a
file, this option is ignored, because the language is obtained from the
extension of the file}

\item{tabSize}{number of spaces of the indentation (usually \code{2} or
\code{4});
if \code{NULL} (the default), there are two possibilities:
if the contents is read from the current file in RStudio, then the
number of spaces will be the one you use in RStudio;
otherwise it is set to \code{2}}
}
\value{
The pretty code in a character string.
}
\description{
Prettify some C, C++ or Java code.
}
\note{
This function requires the command line utility \code{clang-format}.
}
