% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyPYTHON.R
\name{prettifyPython}
\alias{prettifyPython}
\title{Prettify Python}
\usage{
prettifyPython(contents = NA)
}
\arguments{
\item{contents}{the code to be prettified; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}
}
\value{
The pretty code in a character string.
}
\description{
Prettify Python code.
}
\note{
This function requires \href{https://github.com/psf/black}{black}.
}
