% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyXML.R
\name{prettifyXML}
\alias{prettifyXML}
\title{Prettify XML}
\usage{
prettifyXML(contents = NA, tabSize = NULL)
}
\arguments{
\item{contents}{the code to be prettified; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}

\item{tabSize}{number of spaces of the indentation (usually \code{2} or
\code{4});
if \code{NULL} (the default), there are two possibilities:
if the contents are read from the current file in RStudio, then the
number of spaces will be the one you use in RStudio;
otherwise it is set to \code{2}}
}
\value{
The pretty code in a character string.
}
\description{
Prettify some XML or SVG code.
}
\details{
The code is prettified with the help of the command line utility
  \code{xmllint} if it is available, otherwise the \code{xml2} is used.
}
