% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy.R
\name{reindent_bigHTML}
\alias{reindent_bigHTML}
\title{Indent big HTML file}
\usage{
reindent_bigHTML(contents = NA, tabSize = NULL)
}
\arguments{
\item{contents}{the code to be prettified; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}

\item{tabSize}{number of spaces of the indentation (usually \code{2} or
\code{4});
if \code{NULL} (the default), there are two possibilities:
if the contents is read from the current file in RStudio, then the
number of spaces will be the one you use in RStudio;
otherwise it is set to \code{2}}
}
\value{
The indented code in a character string.
}
\description{
Indent a big HTML file.
}
\details{
This function calls \code{\link{tidyHTML}} followed by
  \code{\link{foldCode}}.
}
