% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_V8.R
\name{wordWrap}
\alias{wordWrap}
\title{Word wrap using V8}
\usage{
wordWrap(contents = NA, ncharacters = 80)
}
\arguments{
\item{contents}{the text to be wrapped; there are three possibilities for
this argument:
\code{NA} (default), to use the file currently opened in RStudio;
the path to a file;
or the code given as a character vector}

\item{ncharacters}{target number of characters per line}
}
\value{
The wrapped text in a character string.
}
\description{
Word wrap a text.
}
