% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyAddins-package.R
\docType{package}
\name{prettifyAddins}
\alias{prettifyAddins}
\title{Prettify Addins}
\description{
This package provides some RStudio addins: \code{Prettify} addins and
  \code{Indent} addins. To run an addin, select it from the Addins menu
  within RStudio. The \code{Indent} addins only reindent the code, while
  the \code{Prettify} addins also modify the code, e.g. they add trailing
  semi-colons to JavaScript code when they are missing.
}
\examples{
# Here is the list of supported languages:
list(
  Shiny = list(
    Prettify = c("HTML", "CSS", "SCSS", "JavaScript", "JSX", "Markdown"),
    Indent = c("HTML", "CSS", "SCSS", "JavaScript", "JSX", "C(++)", "Java",
               "Python", "SAS", "Shell", "SQL")
  ),
  V8 = list(
    Prettify = c("HTML", "CSS", "SCSS", "JavaScript", "JSX", "Markdown"),
    Indent = c("HTML", "CSS", "SCSS", "JavaScript", "JSX")
  ),
  CLANG = c("C(++)", "Java"),
  XML = c("XML", "SVG")
)
}
