% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyAddins-package.R
\docType{package}
\name{prettifyAddins}
\alias{prettifyAddins}
\title{Prettify Addins}
\description{
This package provides some RStudio addins: \code{Prettify} addins and
  \code{Indent} addins. To run an addin, select it from the Addins menu
  within RStudio. The \code{Indent} addins only reindent the code, while
  the \code{Prettify} addins also modify the code, e.g. they add trailing
  semi-colons to JavaScript code when they are missing.
}
\examples{
# get the list of supported languages:
getPrettifiableLanguages()
}
