% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyAddins-package.R
\docType{package}
\name{prettifyAddins}
\alias{prettifyAddins}
\title{Prettify Addins}
\description{
This package provides two RStudio addins: \code{Prettify} and \code{Indent}.
To run an addin, select \code{Prettify} or \code{Indent} from the Addins
menu within RStudio. The \code{Indent} addin only reindents the code, while
the \code{Prettify} addin also modifies the code, e.g. it adds trailing
semi-colons to JavaScript code when they are missing.
}
